Abstract store more
