+++ crm-vue/.vscode/settings.json
@@ -10,6 +10,7 @@
     "msapplication",
     "scripthost",
     "sweetalert",
+    "userlists",
     "vite",
     "vuex"
   ]
