+++ crm-vue/src/shims-vuex.d.ts
@@ -1,5 +1,5 @@
 
-import { IRootState } from "./store/interfaces";
+import { IRootState } from "./store/register";
 declare module "@vue/runtime-core" {
     interface ComponentCustomProperties {
         $store: Store<IRootState>;
