+++ crm-vue/src/store/modules/root/getters.ts
@@ -1,5 +1,6 @@
+import { IRootGettersTypes } from "@/store/store_interfaces/root_store_interface";
 import { GetterTree } from "vuex";
-import { IRootGettersTypes, IRootState, IUserData } from "./../../interfaces";
+import { IRootState, IUserData } from "@/store/register";
 
 export const getters: GetterTree<IRootState, IRootState> & IRootGettersTypes = {
   getVersion: (state: IRootState): string => {
