+++ crm-vue/src/store/modules/root/state.ts
@@ -1,4 +1,4 @@
-import { IRootState } from "./../../interfaces";
+import { IRootState } from "../../register";
 
 export const state: IRootState = {
   root: true,
