+++ crm-vue/src/store/modules/task/getters.ts
@@ -1,10 +1,10 @@
 import { Task } from '@/interfaces/Task';
 import { GetterTree } from "vuex";
 import {
-  TaskGettersTypes,
-  TaskStateTypes,
+
   IRootState
-} from "@/store/interfaces";
+} from "@/store/register";
+import { TaskGettersTypes, TaskStateTypes } from '@/store/store_interfaces/task_store_interface';
 export const getters: GetterTree<TaskStateTypes, IRootState> &
   TaskGettersTypes = {
   getAllTasks(state: TaskStateTypes): Task[] {
