+++ crm-vue/src/store/modules/task/index.ts
@@ -1,9 +1,10 @@
 import { Module } from "vuex";
-import { TaskStateTypes, IRootState } from "@/store/interfaces";
+import { IRootState } from "@/store/register";
 import { getters } from "./getters";
 import { actions } from "./actions";
 import { mutations } from "./mutations";
 import { state } from "./state";
+import { TaskStateTypes } from "@/store/store_interfaces/task_store_interface";
 
 // Module
 const task: Module<TaskStateTypes, IRootState> = {
