+++ crm-vue/src/store/modules/task/state.ts
@@ -1,4 +1,5 @@
-import { TaskStateTypes } from "@/store/interfaces";
+import { TaskStateTypes } from "@/store/store_interfaces/task_store_interface";
+
 
 export const state: TaskStateTypes = {
   tasks: null,
