+++ crm-vue/src/store/modules/task/types.ts
@@ -3,7 +3,7 @@ import {
   TaskMutationsTypes,
   TaskGettersTypes,
   TaskActionsTypes
-} from "@/store/interfaces";
+} from "@/store/store_interfaces/task_store_interface";
 import { Store as VuexStore, CommitOptions, DispatchOptions } from "vuex";
 
 export type DailyTaskStoreModuleTypes<S = TaskStateTypes> = Omit<
